a2a/mathAgent/pyproject.toml
[project]
name = "a2a-math-agent-langgraph"
version = "0.1.0"
description = "Math Agent for A2A MCP System"
authors = [
    {name = "A2A MCP Team"}
]
dependencies = [
    "a2a-sdk>=0.2.6,<0.3.0",
    "httpx>=0.25.0",
    "uvicorn>=0.23.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "sympy>=1.12",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "mcp>=1.0.0"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
] 